Move integration tests out of rpc feature project to integration test project
