+++ src/Stratis.Bitcoin.IntegrationTests/RPC/BaseRPCControllerTest.cs
@@ -3,11 +3,10 @@ using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Features.BlockStore;
 using Stratis.Bitcoin.Features.Consensus;
 using Stratis.Bitcoin.Features.MemoryPool;
-using Stratis.Bitcoin.Features.Miner;
-using Stratis.Bitcoin.Features.Wallet;
+using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Tests;
 
-namespace Stratis.Bitcoin.Features.RPC.Tests.Controller
+namespace Stratis.Bitcoin.IntegrationTests.RPC
 {
     /// <summary>
     /// Base class for RPC tests.
